Fix duration paper

See merge request guardian-exam-online/geo-dotnet-api!170

+++ src/Core/Application/Examination/Papers/Dtos/PaperDto.cs
@@ -12,7 +12,7 @@ public class PaperDto : IDto
     public string ExamName { get; set; }
     public int NumberOfQuestion { get; set; }
     public int TotalAttended { get; set; }
-    public int? Duration { get; set; }
+    public float? Duration { get; set; }
     public DateTime? StartTime { get; set; }
     public DateTime? EndTime { get; set; }
     public ShowResult ShowMarkResult { get; set; }
